# Add circular buffer implementation to be later used inside `OrderingSender`