--- a/ipa-core/src/helpers/buffers/mod.rs
+++ b/ipa-core/src/helpers/buffers/mod.rs
@@ -1,6 +1,9 @@
 mod ordering_sender;
 mod unordered_receiver;
 
+#[allow(dead_code)]
+mod circular;
+
 pub use ordering_sender::OrderingSender;
 pub use unordered_receiver::{
     DeserializeError, EndOfStreamError, Error as UnorderedReceiverError, UnorderedReceiver,